docker/backend/packages/ios-device-detector/pyproject.toml
[project]
name = "ios-device-detector"
version = "1.0.0"
description = "Detect iOS device models from image dimensions and metadata"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
image = [
    "pillow>=10.0.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
client = [
    "httpx>=0.24.0",
]
full = [
    "ios-device-detector[image,web,client]",
]
dev = [
    "ios-device-detector[full]",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ios_device_detector"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
